Add ft_intmaxtoa && ft_intmaxtoa_base. ft_vsnprintf now works for signed ints with j modifier

+++ src/ft_digitcount.c
@@ -6,13 +6,14 @@
 /*   By: rhallste <rhallste@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/03 13:56:51 by rhallste          #+#    #+#             */
-/*   Updated: 2017/12/05 20:16:36 by rhallste         ###   ########.fr       */
+/*   Updated: 2017/12/06 17:44:30 by rhallste         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stdint.h>
 #include "../inc/libft.h"
 
-int	ft_digitcount(unsigned long long n)
+int	ft_digitcount(uintmax_t n)
 {
 	return (ft_digitcount_base(n, 10));
 }
